# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * myaiteacher API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  fullName: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface UserProfile {
  id: number;
  username: string;
  fullName: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface SubjectProgress {
  id: number;
  subject: string;
  completedLessons: number;
  totalLessons: number;
  averageScore: number;
  progressPercent: number;
}

export interface LessonActivity {
  id: number;
  subject: string;
  lesson: string;
  score: number;
  status: string;
  createdAt: string;
}

export interface DashboardStats {
  completedLessons: number;
  averageScore: number;
  pendingHomework: number;
  overallProgress: number;
}

export interface DashboardData {
  stats: DashboardStats;
  subjects: SubjectProgress[];
  recentActivity: LessonActivity[];
}

export interface SubjectListItem {
  id: number;
  name: string;
  grade: string;
  description: string;
}

export interface LessonItem {
  id: number;
  lesson: string;
  score: number;
  status: string;
  createdAt: string;
}

export interface SubjectDetail {
  id: number;
  name: string;
  grade: string;
  description: string;
  progressPercent: number;
  completedLessons: number;
  totalLessons: number;
  averageScore: number;
  lessons: LessonItem[];
}

export interface StartLessonInput {
  lesson: string;
  score?: number;
  status?: string;
}

export type KnowledgeTopicMasteryLevel = typeof KnowledgeTopicMasteryLevel[keyof typeof KnowledgeTopicMasteryLevel];


export const KnowledgeTopicMasteryLevel = {
  mastered: 'mastered',
  weak: 'weak',
  not_started: 'not_started',
} as const;

export interface KnowledgeTopic {
  id: number;
  topicName: string;
  score: number;
  status: string;
  masteryLevel: KnowledgeTopicMasteryLevel;
}

export type AIRecommendationType = typeof AIRecommendationType[keyof typeof AIRecommendationType];


export const AIRecommendationType = {
  start: 'start',
  review: 'review',
  repeat: 'repeat',
} as const;

export interface AIRecommendation {
  type: AIRecommendationType;
  message: string;
  topicName: string;
}

export interface KnowledgeTreeData {
  subjectId: number;
  subjectName: string;
  topics: KnowledgeTopic[];
  recommendations: AIRecommendation[];
}

export interface LessonPhase {
  phase: number;
  title: string;
  duration: string;
  description: string;
  activities: string[];
}

export interface BloomLevel {
  level: number;
  name: string;
  color: string;
  description: string;
}

export interface LessonSession {
  id: number;
  lessonId: number;
  currentPhase: number;
  status: string;
  startedAt: string;
  completedAt?: string | null;
}

export interface LessonDetail {
  id: number;
  subjectId: number;
  subjectName: string;
  title: string;
  description: string;
  bloomLevel: number;
  bloomLevels: BloomLevel[];
  phases: LessonPhase[];
  currentSession: LessonSession | null;
}

export interface StartLessonSessionInput {
  lessonId: number;
}

export interface ChatInput {
  message: string;
  lessonId?: number;
}

export interface ChatResponse {
  response: string;
  messageId: number;
}

export interface ChatMessageItem {
  id: number;
  role: string;
  content: string;
  createdAt: string;
}

export type GetChatHistoryParams = {
lessonId?: number;
};