# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * myaiteacher API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  response: string;
  messageId: number;
}