# lib/api-zod/src/generated/types/lessonDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * myaiteacher API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BloomLevel } from './bloomLevel';
import type { LessonPhase } from './lessonPhase';
import type { LessonSession } from './lessonSession';

export interface LessonDetail {
  id: number;
  subjectId: number;
  subjectName: string;
  title: string;
  description: string;
  bloomLevel: number;
  bloomLevels: BloomLevel[];
  phases: LessonPhase[];
  currentSession: LessonSession | null;
}