# lib/api-zod/src/generated/types/lessonSession.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * myaiteacher API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LessonSession {
  id: number;
  lessonId: number;
  currentPhase: number;
  status: string;
  startedAt: Date;
  completedAt?: Date | null;
}